archive/ask-legacy/requirements.txt
# Core dependencies
python-dotenv>=1.0.0  # Environment variable management
requests>=2.31.0  # HTTP library for API calls

# Groq API
groq>=0.4.0  # Groq API client

# API Server dependencies
fastapi>=0.104.0  # FastAPI web framework
uvicorn[standard]>=0.24.0  # ASGI server
python-multipart>=0.0.6  # Form data parsing
pydantic>=2.0.0  # Data validation
pydantic-settings>=2.0.0  # Settings management

# Database
sqlalchemy>=2.0.0  # SQL ORM
psycopg2-binary>=2.9.0  # PostgreSQL adapter
alembic>=1.13.0  # Database migrations (optional)

# Payments
razorpay>=1.4.0  # Razorpay payment gateway

# Authentication
pyjwt>=2.8.0  # JWT token handling
python-jose[cryptography]>=3.3.0  # JWT utilities

# Shared backend package (monorepo)
-e ../../packages/shared-backend  # Local editable install